% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cityList}
\alias{cityList}
\title{Prices of 50,000 round cut diamonds.}
\format{
a data frame containing 209579 obs. of  6 variables
\describe{
\item{id}{unique identifier}
\item{name}{location name}
\item{state}{location state code}
\item{country}{location country code}
\item{coord.lon}{longitude}
\item{coord.lat}{latitude}
}
}
\source{
\url{http://bulk.openweathermap.org/sample/city.list.json.gz}
}
\usage{
cityList
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
