% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagleIO.R
\name{EIO_Node}
\alias{EIO_Node}
\title{Functions for extracting node data from WQI's eagle.IO instance}
\usage{
EIO_Node(APIKEY, param)
}
\arguments{
\item{APIKEY}{== This is required for all functions and can be generated from the account settings in WQI's Eagle.IO instance}

\item{param}{== a node ID of a given parameter in eagle.IO, noting historic will only work with nodes that contain historic data}
}
\value{
data frame containing returned node data
}
\description{
Functions for extracting node data from WQI's eagle.IO instance
}
