% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labRef}
\alias{labRef}
\title{Reference table for all of WQI's location sources with LAB data
Provides Node ID's for available discrete analytes}
\format{
a data frame containing 134 obs. of  56 variables
\describe{
\item{_id}{_id}
\item{GSnum}{GSnum}
\item{name}{name}
\item{TextFileSource}{TextFileSource}
\item{WQ ANALYSIS NO}{WQ ANALYSIS NO}
\item{Site Name}{Site Name}
\item{SITE NO}{SITE NO}
\item{Atrazine}{Atrazine}
\item{Ametryn}{Ametryn}
\item{Ammonium nitrogen}{Ammonium nitrogen}
\item{RT Level}{RT Level}
\item{Bromacil}{Bromacil}
\item{Chlorpyrifos}{Chlorpyrifos}
\item{Conductivity}{Conductivity}
\item{Diss Kjeldahl nitrogen}{Diss Kjeldahl nitrogen}
\item{Diss Kjeldahl phosphor}{Diss Kjeldahl phosphor}
\item{Diss organic nitrogen}{Diss organic nitrogen}
\item{Diss organic phosphor}{Diss organic phosphor}
\item{Diuron}{Diuron}
\item{Filt react phosphorus}{Filt react phosphorus}
\item{Fipronil}{Fipronil}
\item{Fluometuron}{Fluometuron}
\item{Fluroxypyr}{Fluroxypyr}
\item{Glyphosate}{Glyphosate}
\item{Haloxyfop acid}{Haloxyfop acid}
\item{Hexazinone}{Hexazinone}
\item{Imazapic}{Imazapic}
\item{Imidacloprid}{Imidacloprid}
\item{Isoxaflutole}{Isoxaflutole}
\item{MCPA}{MCPA}
\item{Metolachlor}{Metolachlor}
\item{Metribuzin}{Metribuzin}
\item{Metsulfuron-methyl}{Metsulfuron-methyl}
\item{Oxidised nitrogen}{Oxidised nitrogen}
\item{Particulate Nitrogen}{Particulate Nitrogen}
\item{Particulate phosphorus}{Particulate phosphorus}
\item{Pendimethalin}{Pendimethalin}
\item{Prometryn}{Prometryn}
\item{Simazine}{Simazine}
\item{Tebuthiuron}{Tebuthiuron}
\item{Terbuthylazine}{Terbuthylazine}
\item{Terbutryn}{Terbutryn}
\item{Total organic carbon}{Total organic carbon}
\item{Triclopyr}{Triclopyr}
\item{TSS}{TSS}
\item{Ttl Kjeldahl nitrogen}{Ttl Kjeldahl nitrogen}
\item{Ttl nitrogen}{Ttl nitrogen}
\item{Ttl phosphorus}{Ttl phosphorus}
\item{Turbidity}{Turbidity}
\item{Chlorothalonil}{Chlorothalonil}
\item{Diazinon}{Diazinon}
\item{Propiconazole}{Propiconazole}
\item{RT TSSeq}{RT TSSeq}
\item{RT Conductivity}{RT Conductivity}
\item{RT Turbidity}{RT Turbidity}
\item{WQI Site Name}{WQI Site Name}
}
}
\source{
\url{https://wqi.eagle.io/}
}
\usage{
labRef
}
\description{
Reference table for all of WQI's location sources with LAB data
Provides Node ID's for available discrete analytes
}
\keyword{datasets}
