% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loggerRef}
\alias{loggerRef}
\title{Reference table for all of WQI's location sources with Logger data}
\format{
a data frame containing 134 obs. of  56 variables
\describe{
\item{_id}
\item{GSnum}
\item{name}
\item{loggerID}
\item{Main - Level}
\item{OPUSResults - OPUS1000 N NO3}
\item{OPUSResults - OPUS1036 Abs210}
\item{Nitrate - NICO RefA}
\item{OPUSResults - OPUS1042 Abs254}
\item{Nitrate - NICO RefB}
\item{OPUSResults - OPUS1034 Abs360}
\item{Nitrate - NICO RefC}
\item{Nitrate - NICO RefD}
\item{OPUSResults - OPUS1060 SQI}
\item{OPUSResults - OPUS1016 TSSeq}
\item{Public - RSSI}
\item{Resets - WDE}
\item{Public - OPUS Serial}
\item{Public - SondePresent}
\item{Public - NitratePresent}
\item{Public - ADCPPresent}
\item{Main - LoggerBattV Min}
\item{Main - PestFrV Min}
\item{Main - PumpBattV Min}
\item{Public - SamplerOnOff}
\item{Public - SamplerState}
\item{FridgeStatus - NFridgeStatus}
\item{FridgeStatus - PFridgeStatus}
\item{Public - SampleNo}
\item{Sonde - SondeTemp}
\item{Sonde - Conductivity}
\item{Sonde - Turbidity}
\item{Sonde - DO-mgL}
\item{Sonde - fDOM-QSU}

}
}
\source{
\url{https://wqi.eagle.io/}
}
\usage{
loggerRef
}
\description{
Provides Node ID's for available real time parameters
}
\keyword{datasets}
