% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagleIO.R
\name{NICO_Hist}
\alias{NICO_Hist}
\title{Extract historic NICO data from WQI's eagle.IO instance}
\source{
\url{https://wqi.eagle.io/}
}
\usage{
NICO_Hist(
  APIKEY,
  NNO3,
  refA,
  refB,
  refC,
  refD,
  SQI,
  START,
  END = format(Sys.time() + 86400, "\%Y-\%m-\%dT\%H:\%M:\%SZ")
)
}
\arguments{
\item{APIKEY}{This is required for all functions and can be generated from the account settings in WQI's Eagle.IO instance}

\item{NNO3}{A node ID for a given sites NNO3 parameter in eagle.IO.}

\item{refA}{A node ID for a given sites refA parameter in eagle.IO.}

\item{refB}{A node ID for a given sites refB parameter in eagle.IO.}

\item{refC}{A node ID for a given sites refC parameter in eagle.IO.}

\item{refD}{A node ID for a given sites refD parameter in eagle.IO.}

\item{SQI}{A node ID for a given sites SQI parameter in eagle.IO.}

\item{START}{DateTime to start the historic period of capture. Format needs to be in ISO8601, see: 2014-10-09T22:38:10Z || 2014-10-09T22:38:10.000Z || 2014-10-09T20:38:10+0200 || 2014-10-09T20:38:10+02:00}

\item{END}{DateTime to end the historic period of capture. Defaults to the current system time + one day. Format needs to be in ISO8601, see: 2014-10-09T22:38:10Z || 2014-10-09T22:38:10.000Z || 2014-10-09T20:38:10+0200 || 2014-10-09T20:38:10+02:00}
}
\value{
tibble containing returned historic data for OPUS reportable parameters and associsated spectral values, value and quality. Time as "Australia/Brisbane"
}
\description{
Function for extracting a tibble of historic NICO data from WQI's eagle.IO instance.
}
\examples{
loggerRef <- WQI::loggerRef
site <- "1350053"
dex <- loggerRef \%>\%
  dplyr::filter(GSnum == site)

NNO3 <- dex$`OPUSResults - OPUS1000 N NO3`
refA <- dex$`Nitrate - NICO RefA`
refB <- dex$`Nitrate - NICO RefB`
refC <- dex$`Nitrate - NICO RefC`
refD <- dex$`Nitrate - NICO RefD`
SQI <- dex$`OPUSResults - OPUS1060 SQI`

}
