% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenWeather.R
\name{OpenWeather}
\alias{OpenWeather}
\title{Access to the OpenWeather API data (Free Tier)  --- https://openweathermap.org/
example API :: https://api.openweathermap.org/data/2.5/onecall/timemachine?lat={lat}&lon={lon}&dt={time}&appid={API key}
hourly Hourly forecast weather data API response
hourly.dt Time of the forecasted data, Unix, UTC
hourly.temp Temperature. Units – default: kelvin, metric: Celsius, imperial: Fahrenheit. How to change units used
hourly.feels_like Temperature. This accounts for the human perception of weather. Units – default: kelvin, metric: Celsius, imperial: Fahrenheit.
hourly.pressure Atmospheric pressure on the sea level, hPa
hourly.humidity Humidity, \%
hourly.dew_point Atmospheric temperature (varying according to pressure and humidity) below which water droplets begin to condense and dew can form. Units – default: kelvin, metric: Celsius, imperial: Fahrenheit.
hourly.uvi UV index
hourly.clouds Cloudiness, \%
hourly.visibility Average visibility, metres
hourly.wind_speed Wind speed. Units – default: metre/sec, metric: metre/sec, imperial: miles/hour.How to change units used
hourly.wind_gust (where available) Wind gust. Units – default: metre/sec, metric: metre/sec, imperial: miles/hour. How to change units used
chourly.wind_deg Wind direction, degrees (meteorological)
hourly.pop Probability of precipitation
hourly.rain
hourly.rain.1h (where available) Rain volume for last hour, mm
hourly.snow
hourly.snow.1h (where available) Snow volume for last hour, mm
hourly.weather
hourly.weather.id Weather condition id
hourly.weather.main Group of weather parameters (Rain, Snow, Extreme etc.)
hourly.weather.description Weather condition within the group (full list of weather conditions). Get the output in your language
hourly.weather.icon Weather icon id. How to get icons}
\usage{
OpenWeather(City, time, APIKEY)
}
\arguments{
\item{City}{== the City as displayed in the cityList file extracted from the website}

\item{time}{== UNIX time for a given day. Free tier only supports the last 5 days
importFrom magrittr "\%>\%"}

\item{APIKEY}{== given from the free tier account for OpenWeather}
}
\value{
data frame containing weather observation data from given city. Time as "Australia/Brisbane" (UTC+10)
}
\description{
Access to the OpenWeather API data (Free Tier)  --- https://openweathermap.org/
example API :: https://api.openweathermap.org/data/2.5/onecall/timemachine?lat={lat}&lon={lon}&dt={time}&appid={API key}
hourly Hourly forecast weather data API response
hourly.dt Time of the forecasted data, Unix, UTC
hourly.temp Temperature. Units – default: kelvin, metric: Celsius, imperial: Fahrenheit. How to change units used
hourly.feels_like Temperature. This accounts for the human perception of weather. Units – default: kelvin, metric: Celsius, imperial: Fahrenheit.
hourly.pressure Atmospheric pressure on the sea level, hPa
hourly.humidity Humidity, \%
hourly.dew_point Atmospheric temperature (varying according to pressure and humidity) below which water droplets begin to condense and dew can form. Units – default: kelvin, metric: Celsius, imperial: Fahrenheit.
hourly.uvi UV index
hourly.clouds Cloudiness, \%
hourly.visibility Average visibility, metres
hourly.wind_speed Wind speed. Units – default: metre/sec, metric: metre/sec, imperial: miles/hour.How to change units used
hourly.wind_gust (where available) Wind gust. Units – default: metre/sec, metric: metre/sec, imperial: miles/hour. How to change units used
chourly.wind_deg Wind direction, degrees (meteorological)
hourly.pop Probability of precipitation
hourly.rain
hourly.rain.1h (where available) Rain volume for last hour, mm
hourly.snow
hourly.snow.1h (where available) Snow volume for last hour, mm
hourly.weather
hourly.weather.id Weather condition id
hourly.weather.main Group of weather parameters (Rain, Snow, Extreme etc.)
hourly.weather.description Weather condition within the group (full list of weather conditions). Get the output in your language
hourly.weather.icon Weather icon id. How to get icons
}
