% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagleIO.R
\name{EIO_Comms}
\alias{EIO_Comms}
\title{Extract communications information for a given logger source}
\source{
url{https://wqi.eagle.io/}
}
\usage{
EIO_Comms(APIKEY, param)
}
\arguments{
\item{APIKEY}{== This is required for all functions and can be generated from the account settings in WQI's Eagle.IO instance}

\item{param}{== a node ID of a given parameter in eagle.IO, noting historic will only work with nodes that contain historic data}
}
\value{
data frame containing returned comms data
}
\description{
Function for extracting comms data from WQI's eagle.IO instance. Requires a node ID for a DATA SOURCE that has connection status information to return (ie. Campbell pakbus logger)
}
