% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagleIO.R
\name{EIO_Hist}
\alias{EIO_Hist}
\title{Functions for extracting historic data from WQI's eagle.IO instance}
\usage{
EIO_Hist(APIKEY, param, START)
}
\arguments{
\item{APIKEY}{== This is required for all functions and can be generated from the account settings in WQI's Eagle.IO instance}

\item{param}{== a node ID of a given parameter in eagle.IO, noting historic will only work with nodes that contain historic data}

\item{START}{== number of days to lookback}
}
\value{
data frame containing returned historic data
}
\description{
Functions for extracting historic data from WQI's eagle.IO instance
}
