% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eagleIO.R
\name{EIO_Node}
\alias{EIO_Node}
\title{Extract node data from WQI's eagle.IO instance}
\source{
url{https://wqi.eagle.io/}
}
\usage{
EIO_Node(APIKEY, param)
}
\arguments{
\item{APIKEY}{== This is required for all functions and can be generated from the account settings in WQI's Eagle.IO instance}

\item{param}{== a node ID of a given parameter in eagle.IO, noting historic will only work with nodes that contain historic data}
}
\value{
data frame containing returned node data
}
\description{
Will extract the current value for a given NODE ID from WQI's eagle.IO instance.
}
\examples{
#content <- EIO_Node(APIKEY = "XYZ", param = "59cca1064f2ee90c99b94b2e")

}
