% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labRef}
\alias{labRef}
\title{Reference table for all of WQI's location sources with LAB data}
\format{
a data frame containing 134 obs. of  56 variables
\describe{
\item{_id}
\item{GSnum}
\item{name}
\item{TextFileSource}
\item{WQ ANALYSIS NO}
\item{Site Name}
\item{SITE NO}
\item{Atrazine}
\item{Ametryn}
\item{Ammonium nitrogen}
\item{RT Level}
\item{Bromacil}
\item{Chlorpyrifos}
\item{Conductivity}
\item{Diss Kjeldahl nitrogen}
\item{Diss Kjeldahl phosphor}
\item{Diss organic nitrogen}
\item{Diss organic phosphor}
\item{Diuron}
\item{Filt react phosphorus}
\item{Fipronil}
\item{Fluometuron}
\item{Fluroxypyr}
\item{Glyphosate}
\item{Haloxyfop acid}
\item{Hexazinone}
\item{Imazapic}
\item{Imidacloprid}
\item{Isoxaflutole}
\item{MCPA}
\item{Metolachlor}
\item{Metribuzin}
\item{Metsulfuron-methyl}
\item{Oxidised nitrogen}
\item{Particulate Nitrogen}
\item{Particulate phosphorus}
\item{Pendimethalin}
\item{Prometryn}
\item{Simazine}
\item{Tebuthiuron}
\item{Terbuthylazine}
\item{Terbutryn}
\item{Total organic carbon}
\item{Triclopyr}
\item{TSS}
\item{Ttl Kjeldahl nitrogen}
\item{Ttl nitrogen}
\item{Ttl phosphorus}
\item{Turbidity}
\item{Chlorothalonil}
\item{Diazinon}
\item{Propiconazole}
\item{RT TSSeq}
\item{RT Conductivity}
\item{RT Turbidity}
\item{WQI Site Name}

}
}
\source{
\url{https://wqi.eagle.io/}
}
\usage{
labRef
}
\description{
Provides Node ID's for available discrete analytes
}
\keyword{datasets}
