% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WMIP.R
\name{WMIP_Extract}
\alias{WMIP_Extract}
\title{Extracts data from the Water Monitoring Information Portal (WMIP)
Website: https://water-monitoring.information.qld.gov.au/}
\usage{
WMIP_Extract(WMIPID, Vfrom, Vto, START)
}
\arguments{
\item{WMIPID}{== a Gauging station number}

\item{Vfrom}{== refers to parameter codes (100.00 = level)}

\item{Vto}{== allows you to leverage the platform to convert level to discharge (ie. Vfrom = 100.00 \link{level} , Vto = 140.00 \link{discharge})}

\item{START}{== date* to lookback to. format = YYYYMMDD}
}
\value{
a data frame containing extracted WMIP data for specified gauging station/parameter.
}
\description{
Extracts data from the Water Monitoring Information Portal (WMIP)
Website: https://water-monitoring.information.qld.gov.au/
}
